\name{bisoreg-package}
\alias{bisoreg-package}
\docType{package}

\title{
Bayesian Isotonic Regression with Bernstein Polynomials
}

\description{
  The function \code{bisoreg} implements the procedure described in the paper "A Bayesian variable selection approach to monotonic regression." The package also contains additional functions to compare with the Bayesian method in the paper.
}

\details{
  \tabular{ll}{
    Package: \tab bisoreg\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2008-10-07\cr
    License: \tab GPL (>=2)\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
S. McKay Curtis (with helpful suggestions from Sujit K. Ghosh)

Maintainer: S. McKay Curtis
}

\references{
Curtis, S. M. and Ghosh, S. K. (2009) "A variable selection approach to monotonic regression with Bernstein polynomials." (unpublished manuscript).
}

\examples{
## See examples for bisoreg ##
}

\keyword{ package }
