\name{childgrowth}
\docType{data}
\alias{childgrowth}
\title{Heights of a ten-year-old boy over the course of a year}
\description{This data set is the \code{onechild} dataset from the \code{fda} package.
}
\usage{childgrowth}
\format{A \code{data.frame} containing 83 observations on two variables.

\code{day} is day of the year on which the height measurement was taken.

\code{height} is the height in inches of the child.
}
\source{See the \code{fda} package.}
\references{
Ramsay, James O., and Silverman, Bernard W. (2002). \emph{Applied Functional Data Analysis 2nd ed.} Springer, New York.

Tuddenham, R. D., and Snyder, M. M. (1954). "Physical growth of California boys and girls from birth to age 18", \emph{University of California Publications in Child Development}, 1, 183-364.
}
\keyword{datasets}