\name{fits.isoreg}
\alias{fits.isoreg}
\title{ Computes fitted values for isoreg }
\description{
Computes the fitted values for an object that has been creatd using the
isoreg function
}
\usage{
fits.isoreg(iso, x0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{iso}{ on object created with the function \code{isoreg} }
  \item{x0}{ x-values at which to compute the fitted values }
}
\details{None.}
\value{A vector of fitted values.}
\references{None.}
\author{S. McKay Curtis}
\note{None.}
\seealso{\code{\link{isoreg}}}
\examples{
## None ##
}
\keyword{ nonparametric }
