\name{get.W}
\alias{get.W}
\title{ Monotonic Bernstein Polynomial Basis Expansion}
\description{
Computes the monotonic, Bernstein polynomial basis expansion matrix "W" for a vector x.
}
\usage{
get.W(x, m)
}
\arguments{
  \item{x}{ vector }
  \item{m}{ order of the Bernstein polynomial expansion }
}
\details{None.}
\value{A matrix with dimensions \code{length(x)} by \code{m}.}
\references{None.}
\author{S. McKay Curtis}
\note{None.}
\seealso{\code{\link{bisoreg}}}
\examples{
## See the code to the bisoreg function
}
\keyword{smooth}