\name{print.biso}
\alias{print.biso}

\title{
  Print a biso Object
}

\description{
  Prints a summary of a \code{biso} object.
}

\usage{
\method{print}{biso}(x, digits = 3, ...)
}

\arguments{
  \item{x}{
    a \code{biso} object.
  }
  \item{digits}{
    nonnegative integer specifying the number of digits to round the output.
  }
  \item{\dots}{
    further arguments to pass to \code{print}.
  }
}

\details{
  None.
}

\value{
  None.
}

\references{
  None.
}

\author{
  S. McKay Curtis
}

\note{
  None.
}

\seealso{
  \code{\link{summary.biso}}
}

\examples{
## See examples for bisoreg ##
}

\keyword{ methods }
