\name{as.mcmc.biso}
\alias{as.mcmc.biso}

\title{
  Coerce biso Object to mcmc Object
}

\description{
  Coerces a \code{biso} object to an \code{mcmc} object for analysis with the \code{coda} package.
}

\usage{
\method{as.mcmc}{biso}(x)
}

\arguments{
  \item{x}{
    a \code{biso} object.
  }
}

\details{
  None.
}

\value{
  An S3 object with class \code{mcmc}.
}

\references{
  None.
}

\author{
  S. McKay Curtis
}

\note{
  None.
}

\seealso{
  \pkg{coda} package.
}

\examples{
## None ##
}

\keyword{ methods }
