\name{bisoreg-package}
\alias{bisoreg-package}
\docType{package}

\title{
Bayesian Isotonic Regression with Bernstein Polynomials
}

\description{
  The function \code{bisoreg} implements the procedure described in the paper "A Bayesian variable selection approach to monotonic regression." The package also contains additional functions to compare with the Bayesian method in the paper.
}

\author{
S. McKay Curtis (with helpful suggestions from Sujit K. Ghosh)

Maintainer: S. McKay Curtis
}

\references{
Curtis, S. M. and Ghosh, S. K. (2009). "A variable selection approach to monotonic regression with Bernstein polynomials." Journal of Applied Statistics, 38(5), 961-976.
}

\examples{
## See examples for bisoreg ##
}

\keyword{ package }
